+++ docs/CHANGELOG.md
@@ -44,6 +44,8 @@ be enable with an 2nd Optional Boolean Argument as `true`
 - added function to convert compressed json in pretty json and the other way around
 - added tests for JSON
 - added Tests to the SCRIPT Vars in Lua
+- fixed an Error in Join path function and a build script error
+- fixed secure path for UNIX Systems
 
 Read the DOCS for Version 0.2.0 for more Infos